@@ -3,21 +3,21 @@ description: bottus
 version: 1.0.0
 
 environment:
-  sdk: '>=2.16.1 <3.0.0'
+  sdk: '>=3.2.0'
 
 publish_to: none
 
 dev_dependencies:
-  lints: ^2.0.0
+  lints: ^3.0.0
 dependencies: 
   logging: ^1.0.2
-  nyxx: ^5.0.1
-  nyxx_interactions: ^4.2.0
+  nyxx: ^6.0.0
   modrinth_api:
     git: 
       url: https://github.com/gliscowo/modrinth_api
-      ref: 0497d6f
-  intl: ^0.18.0
+      ref: c1154c8
+  intl: ^0.19.0
   yaml: ^3.1.1
   image: ^4.0.10
-  http: ^0.13.5
+  http: ^1.1.2
+  nyxx_commands: ^6.0.2
