@@ -3,12 +3,11 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:http/http.dart';
-import 'package:logging/logging.dart';
 import 'package:modrinth_api/modrinth_api.dart';
 import 'package:nyxx/nyxx.dart';
-import 'package:nyxx_interactions/nyxx_interactions.dart';
+import 'package:nyxx_commands/nyxx_commands.dart';
 
-import 'announcement_handling.dart' as announcements;
+import 'announcement_handling.dart';
 import 'chat_handlers.dart';
 import 'command_handlers.dart';
 import 'web_data.dart';
@@ -17,7 +16,7 @@ final _logger = Logger("bot");
 final modrinth = ModrinthApi.createClient("wisp-forest/mittwoch-bot");
 final http = Client();
 
-late final INyxxWebsocket bot;
+late final NyxxGateway bot;
 late final Map<String, dynamic> _config;
 
 void main() async {
@@ -29,68 +28,42 @@ void main() async {
   _config = jsonDecode(openConfig("config").readAsStringSync());
 
   final token = jsonDecode(detailsFile.readAsStringSync())["token"] as String;
-  final privilegedGuild = _config["privileged_guild"] as int;
 
-  bot = NyxxFactory.createNyxxWebsocket(token, GatewayIntents.allUnprivileged | GatewayIntents.messageContent,
-      options: ClientOptions(
-          initialPresence: PresenceBuilder.of(
-              status: UserStatus.idle, activity: ActivityBuilder("being a better Mittwoch", ActivityType.competing))
-            ..afk = true));
+  final commands = CommandsPlugin(prefix: null, options: CommandsOptions(type: CommandType.slashOnly));
+  registerCommands(commands);
+  registerAnnouncementCommand(commands);
 
-  bot
-    ..registerPlugin(Logging())
-    ..registerPlugin(CliIntegration())
-    ..registerPlugin(IgnoreExceptions())
-    ..registerPlugin(CloseHttp())
-    ..connect();
+  bot = await Nyxx.connectGatewayWithOptions(
+    GatewayApiOptions(
+      token: token,
+      intents: GatewayIntents.allUnprivileged | GatewayIntents.messageContent,
+      initialPresence: PresenceBuilder(
+        isAfk: true,
+        status: CurrentUserStatus.idle,
+        activities: [ActivityBuilder(name: "being a better Mittwoch", type: ActivityType.competing)],
+      ),
+    ),
+    GatewayClientOptions(plugins: [
+      logging,
+      // cliIntegration,
+      ignoreExceptions,
+      ShutdownHook(),
+      commands,
+    ]),
+  );
 
   await earlyLogging.cancel();
 
-  final interactions = IInteractions.create(WebsocketInteractionBackend(bot))
-    ..registerSlashCommand(SlashCommandBuilder("mittwoch", "Foil", [])..registerHandler(handleMittwochCommand))
-    ..registerSlashCommand(SlashCommandBuilder("faq", "Get the URl to a Wisp Forest FAQ entry", [
-      CommandOptionBuilder(CommandOptionType.string, "entry", "The entry to query", required: true, autoComplete: true)
-        ..autocompleteHandler = autocompleteHandler(faqMappings.keys, "entry")
-    ])
-      ..registerHandler(handleFaqCommand))
-    ..registerSlashCommand(SlashCommandBuilder("announce", "Announces a new mod release", [],
-        guild: Snowflake.value(privilegedGuild), requiredPermissions: PermissionsConstants.administrator)
-      ..registerHandler(announcements.handleAnnounceCommand))
-    ..registerSlashCommand(SlashCommandBuilder("docs", "Get a URL to the specified Wisp Forest docs page", [
-      CommandOptionBuilder(CommandOptionType.string, "path", "The page to query", required: true, autoComplete: true)
-        ..autocompleteHandler = autocompleteHandler(docEntries, "path")
-    ])
-      ..registerHandler(handleDocsCommand))
-    ..registerSlashCommand(SlashCommandBuilder("truth", "Tells you the truth", [])..registerHandler(handleTruth))
-    ..registerSlashCommand(SlashCommandBuilder("lie", "Tells you a lie", [])..registerHandler(handleLie))
-    ..events.onModalEvent.listen(announcements.handleAnnounceModal)
-    ..events.onButtonEvent.listen(announcements.handlePublishCancel)
-    ..events.onButtonEvent.listen(announcements.handleAskForPublish)
-    ..events.onButtonEvent.listen(announcements.handlePublishConfirm)
-    ..syncOnReady();
-
-  stdin.transform(systemEncoding.decoder).transform(LineSplitter()).listen((event) async {
-    if (event == "reset-commands") {
-      _logger.warning("Resetting commands");
-
-      await interactions.deleteGlobalCommands();
-      _logger.info("Global commands deleted");
-
-      await interactions.deleteGuildCommands([Snowflake.value(privilegedGuild)]);
-      _logger.info("Guild commands deleted");
-
-      await interactions.sync();
-      _logger.info("Synced successfully");
-    }
-  });
-
-  bot.eventsWs.onMessageReceived.listen(handleOwl);
-  bot.eventsWs.onMessageReceived.listen(handleTLauncher);
+  bot.onMessageCreate.listen(handleOwl);
+  bot.onMessageCreate.listen(handleTLauncher);
 }
 
-class CloseHttp extends BasePlugin {
+class ShutdownHook extends NyxxPlugin {
   @override
-  FutureOr<void> onBotStop(INyxx nyxx, Logger logger) async => http.close();
+  Future<void> doClose(Nyxx client, Future<void> Function() close) async {
+    http.close();
+    modrinth.dispose();
+  }
 }
 
 File openConfig(String filename) {
@@ -104,6 +77,4 @@ File openConfig(String filename) {
   return file;
 }
 
-Map<String, dynamic> getConfig() {
-  return _config;
-}
+Map<String, dynamic> get botConfig => _config;
